Add files via upload
Here we are discussing about the real time project of analyzing data for establishing new bank .Below are the details of the project 
Neo-Banks are a recent development in the financial sector; they are new banks that solely operate online.
I believed that there should be some kind of connection between the digital world, these new age institutions, and cryptocurrencies.
So I made the decision to start a new project called Data Bank! 
Customers of Data Bank receive cloud data storage allotments that are directly related to the balances in their accounts. The Data Bank team needs your assistance since this business model comes with some intriguing drawbacks.
This case study focuses on metrics calculations, business growth, and smart data analysis to assist the company more accurately estimate and plan for the future.